requirements.txt
dearpygui>=2.0.0
dearpygui_ext>=0.9.5
diffusers>=0.32.0.dev0
icecream>=2.1.3
imageio>=2.22.2
jaxtyping>=0.2.38
kornia>=0.7.0
matplotlib>=3.7.1
mediapy>=1.2.2
numpy>=1.23.3
omegaconf>=2.3.0
onnxruntime>=1.20.1
open3d>=0.17.0
opencv_python>=4.6.0.66
opencv_python_headless>=4.9.0.80
Pillow>=11.1.0
plyfile>=1.1
psutil>=5.9.0
pycocotools>=2.0.5
pyquaternion>=0.9.9
pytest>=7.1.2
PyYAML>=6.0.2
rich>=13.9.4
scikit_image>=0.19.3
scipy>=1.10.0
setuptools>=63.4.1
torch>=2.2.0+cu118
torchvision>=0.17.0+cu118
tqdm>=4.64.1
transformers>=4.29.1
trimesh>=3.21.7
